Storage release

+++ src/main/java/com/FinalTask/entity/Bills.java
@@ -1,6 +1,7 @@
 package com.FinalTask.entity;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 
 public class Bills implements Serializable {
@@ -8,7 +9,7 @@ public class Bills implements Serializable {
     private Date billDate;
     private Users user_name;
     private Bill_status statusId;
-    private Float billValue;
+    private BigDecimal billValue;
 
     public Bills() {
 
@@ -50,11 +51,11 @@ public class Bills implements Serializable {
         this.statusId = statusId;
     }
 
-    public Float getBillValue() {
+    public BigDecimal getBillValue() {
         return billValue;
     }
 
-    public void setBillValue(Float billValue) {
+    public void setBillValue(BigDecimal billValue) {
         this.billValue = billValue;
     }
 }
